mdns: Extract functional tests
Relates to #777.
